cirrus-project/feeders/earthsearch-sentinel-2/requirements.txt
cirrus-geo
pystac-client
stactask